feat: added validations to polls

+++ AssignmentB/src/main/java/no/feedapp/group2/FeedApp/services/PollService.java
@@ -1,8 +1,9 @@
 package no.feedapp.group2.FeedApp.services;
 
-import jakarta.annotation.Nullable;
 import no.feedapp.group2.FeedApp.DTO.Poll.PollCreateDTO;
 import no.feedapp.group2.FeedApp.DTO.Poll.PollUpdateDTO;
+import no.feedapp.group2.FeedApp.controllers.exceptions.CustomerNotFoundException;
+import no.feedapp.group2.FeedApp.controllers.exceptions.PollNotFoundException;
 import no.feedapp.group2.FeedApp.domain.Customer;
 import no.feedapp.group2.FeedApp.domain.Poll;
 import no.feedapp.group2.FeedApp.repositories.CustomerRepository;
@@ -13,7 +14,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class PollService implements IPollService{
+public class PollService implements IPollService {
 
     private final PollRepository pollRepository;
     private final CustomerRepository customerRepository;
@@ -25,30 +26,43 @@ public class PollService implements IPollService{
     }
 
     @Override
-    public Poll createPoll(PollCreateDTO pollCreateDTO) {
+    public Poll createPoll(PollCreateDTO pollCreateDTO) throws CustomerNotFoundException {
         Customer customer = customerRepository.findByUserId(pollCreateDTO.getUserId());
-        Poll poll = PollCreateDTO.convertToPoll(pollCreateDTO,customer);
+        if (customer == null) {
+            throw new CustomerNotFoundException(pollCreateDTO.getUserId());
+        }
+
+        Poll poll = PollCreateDTO.convertToPoll(pollCreateDTO, customer);
         return pollRepository.save(poll);
     }
 
-    @Nullable
     @Override
-    public Poll getPollById(Long id) {
-       return pollRepository.getPollById(id);
+    public Poll getPollById(Long id) throws PollNotFoundException {
+        var poll = pollRepository.getPollById(id);
+        if (poll == null) {
+            throw new PollNotFoundException(id);
+        }
+
+        return poll;
     }
 
 
     @Override
-    public List<Poll> getPollsByUserId(Long userId) {
+    public List<Poll> getPollsByUserId(Long userId) throws CustomerNotFoundException {
+        var customer = customerRepository.findByUserId(userId);
+        if (customer == null) {
+            throw new CustomerNotFoundException(userId);
+        }
+
         return pollRepository.findPollsByUserUserId(userId);
     }
 
-    @Nullable
     @Override
-    public Poll updatePoll(Long id, PollUpdateDTO pollUpdateDTO) {
+    public Poll updatePoll(Long id, PollUpdateDTO pollUpdateDTO) throws PollNotFoundException {
         Poll existingPoll = getPollById(id);
-        if (existingPoll == null){
-            return null;
+
+        if (existingPoll == null) {
+            throw new PollNotFoundException(id);
         }
 
         existingPoll.setQuestion(pollUpdateDTO.getQuestion());
@@ -62,12 +76,17 @@ public class PollService implements IPollService{
     }
 
     @Override
-    public void deletePoll(Long id) {
+    public void deletePoll(Long id) throws PollNotFoundException {
+        var poll = pollRepository.getPollById(id);
+        if (poll == null) throw new PollNotFoundException(id);
         pollRepository.deletePollById(id);
     }
 
     @Override
-    public void deletePollsByUserId(Long userId) {
+    public void deletePollsByUserId(Long userId) throws CustomerNotFoundException {
+        var customer = customerRepository.findByUserId(userId);
+        if (customer == null) throw new CustomerNotFoundException(userId);
+
         pollRepository.deletePollsByUserUserId(userId);
     }
 }
